repo-updater: move Blame and ResolveRevision from vcs/git to gitserver/commands package
