+++ cmd/frontend/backend/inventory.go
@@ -30,7 +30,7 @@ var inventoryCache = rcache.New(fmt.Sprintf("inv:v2:enhanced_%v", useEnhancedLan
 // InventoryContext returns the inventory context for computing the inventory for the repository at
 // the given commit.
 func InventoryContext(repo api.RepoName, db database.DB, commitID api.CommitID, forceEnhancedLanguageDetection bool) (inventory.Context, error) {
-	if !git.IsAbsoluteRevision(string(commitID)) {
+	if !gitserver.IsAbsoluteRevision(string(commitID)) {
 		return inventory.Context{}, errors.Errorf("refusing to compute inventory for non-absolute commit ID %q", commitID)
 	}
