+++ cmd/frontend/graphqlbackend/git_object.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
-	"github.com/sourcegraph/sourcegraph/internal/vcs/git"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
@@ -43,7 +42,7 @@ func (GitObjectID) ImplementsGraphQLType(name string) bool {
 }
 
 func (id *GitObjectID) UnmarshalGraphQL(input any) error {
-	if input, ok := input.(string); ok && git.IsAbsoluteRevision(input) {
+	if input, ok := input.(string); ok && gitserver.IsAbsoluteRevision(input) {
 		*id = GitObjectID(input)
 		return nil
 	}
