+++ cmd/frontend/graphqlbackend/git_revision.go
@@ -5,7 +5,7 @@ import (
 	"net/url"
 	"strings"
 
-	"github.com/sourcegraph/sourcegraph/internal/vcs/git"
+	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 )
 
 type gitRevSpecExpr struct {
@@ -16,7 +16,7 @@ type gitRevSpecExpr struct {
 func (r *gitRevSpecExpr) Expr() string { return r.expr }
 
 func (r *gitRevSpecExpr) Object(ctx context.Context) (*gitObject, error) {
-	oid, err := git.ResolveRevision(ctx, r.repo.db, r.repo.RepoName(), r.expr, git.ResolveRevisionOptions{})
+	oid, err := gitserver.ResolveRevision(ctx, r.repo.db, r.repo.RepoName(), r.expr, gitserver.ResolveRevisionOptions{})
 	if err != nil {
 		return nil, err
 	}
