+++ enterprise/internal/batches/testing/mock_sync_state.go
@@ -10,7 +10,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/gitserver"
 	"github.com/sourcegraph/sourcegraph/internal/repoupdater"
 	"github.com/sourcegraph/sourcegraph/internal/repoupdater/protocol"
-	"github.com/sourcegraph/sourcegraph/internal/vcs/git"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
@@ -20,7 +19,7 @@ type MockedChangesetSyncState struct {
 
 	execReader      func([]string) (io.ReadCloser, error)
 	mockRepoLookup  func(protocol.RepoLookupArgs) (*protocol.RepoLookupResult, error)
-	resolveRevision func(string, git.ResolveRevisionOptions) (api.CommitID, error)
+	resolveRevision func(string, gitserver.ResolveRevisionOptions) (api.CommitID, error)
 }
 
 // MockChangesetSyncState sets up mocks such that invoking SetDerivedState() with
@@ -35,7 +34,7 @@ func MockChangesetSyncState(repo *protocol.RepoInfo) *MockedChangesetSyncState {
 
 		execReader:      gitserver.Mocks.ExecReader,
 		mockRepoLookup:  repoupdater.MockRepoLookup,
-		resolveRevision: git.Mocks.ResolveRevision,
+		resolveRevision: gitserver.Mocks.ResolveRevision,
 	}
 
 	repoupdater.MockRepoLookup = func(args protocol.RepoLookupArgs) (*protocol.RepoLookupResult, error) {
@@ -73,7 +72,7 @@ index 884601b..c4886d5 100644
 		return io.NopCloser(strings.NewReader(testGitHubDiff)), nil
 	}
 
-	git.Mocks.ResolveRevision = func(spec string, opt git.ResolveRevisionOptions) (api.CommitID, error) {
+	gitserver.Mocks.ResolveRevision = func(spec string, opt gitserver.ResolveRevisionOptions) (api.CommitID, error) {
 		return "mockcommitid", nil
 	}
 
@@ -83,6 +82,6 @@ index 884601b..c4886d5 100644
 // Unmock resets the mocks set up by MockGitHubChangesetSync.
 func (state *MockedChangesetSyncState) Unmock() {
 	gitserver.Mocks.ExecReader = state.execReader
-	git.Mocks.ResolveRevision = state.resolveRevision
+	gitserver.Mocks.ResolveRevision = state.resolveRevision
 	repoupdater.MockRepoLookup = state.mockRepoLookup
 }
