+++ internal/gitserver/commands.go
@@ -16,6 +16,9 @@ import (
 	"sync"
 	"time"
 
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
+
 	"github.com/go-git/go-git/v5/plumbing/format/config"
 	"github.com/golang/groupcache/lru"
 	"github.com/opentracing/opentracing-go/log"
@@ -639,4 +642,282 @@ func (c *ClientImplementor) LogReverseEach(repo string, commit string, n int, on
 	defer stdout.Close()
 
 	return errors.Wrap(gitdomain.ParseLogReverseEach(stdout, onLogEntry), "ParseLogReverseEach")
+}
+
+// BlameOptions configures a blame.
+type BlameOptions struct {
+	NewestCommit api.CommitID `json:",omitempty" url:",omitempty"`
+	OldestCommit api.CommitID `json:",omitempty" url:",omitempty"` // or "" for the root commit
+
+	StartLine int `json:",omitempty" url:",omitempty"` // 1-indexed start byte (or 0 for beginning of file)
+	EndLine   int `json:",omitempty" url:",omitempty"` // 1-indexed end byte (or 0 for end of file)
+}
+
+// A Hunk is a contiguous portion of a file associated with a commit.
+type Hunk struct {
+	StartLine int // 1-indexed start line number
+	EndLine   int // 1-indexed end line number
+	StartByte int // 0-indexed start byte position (inclusive)
+	EndByte   int // 0-indexed end byte position (exclusive)
+	api.CommitID
+	Author   gitdomain.Signature
+	Message  string
+	Filename string
+}
+
+// BlameFile returns Git blame information about a file.
+func BlameFile(ctx context.Context, db database.DB, repo api.RepoName, path string, opt *BlameOptions, checker authz.SubRepoPermissionChecker) ([]*Hunk, error) {
+	span, ctx := ot.StartSpanFromContext(ctx, "Git: BlameFile")
+	span.SetTag("repo", repo)
+	span.SetTag("path", path)
+	span.SetTag("opt", opt)
+	defer span.Finish()
+	return blameFileCmd(ctx, gitserverGitCommandFunc(repo, db), path, opt, repo, checker)
+}
+
+func blameFileCmd(ctx context.Context, command gitCommandFunc, path string, opt *BlameOptions, repo api.RepoName, checker authz.SubRepoPermissionChecker) ([]*Hunk, error) {
+	a := actor.FromContext(ctx)
+	if hasAccess, err := authz.FilterActorPath(ctx, checker, a, repo, path); err != nil || !hasAccess {
+		return nil, err
+	}
+	if opt == nil {
+		opt = &BlameOptions{}
+	}
+	if opt.OldestCommit != "" {
+		return nil, errors.Errorf("OldestCommit not implemented")
+	}
+	if err := checkSpecArgSafety(string(opt.NewestCommit)); err != nil {
+		return nil, err
+	}
+	if err := checkSpecArgSafety(string(opt.OldestCommit)); err != nil {
+		return nil, err
+	}
+
+	args := []string{"blame", "-w", "--porcelain"}
+	if opt.StartLine != 0 || opt.EndLine != 0 {
+		args = append(args, fmt.Sprintf("-L%d,%d", opt.StartLine, opt.EndLine))
+	}
+	args = append(args, string(opt.NewestCommit), "--", filepath.ToSlash(path))
+
+	out, err := command(args).Output(ctx)
+	if err != nil {
+		return nil, errors.WithMessage(err, fmt.Sprintf("git command %v failed (output: %q)", args, out))
+	}
+	if len(out) == 0 {
+		return nil, nil
+	}
+
+	commits := make(map[string]gitdomain.Commit)
+	filenames := make(map[string]string)
+	hunks := make([]*Hunk, 0)
+	remainingLines := strings.Split(string(out[:len(out)-1]), "\n")
+	byteOffset := 0
+	for len(remainingLines) > 0 {
+		// Consume hunk
+		hunkHeader := strings.Split(remainingLines[0], " ")
+		if len(hunkHeader) != 4 {
+			return nil, errors.Errorf("Expected at least 4 parts to hunkHeader, but got: '%s'", hunkHeader)
+		}
+		commitID := hunkHeader[0]
+		lineNoCur, _ := strconv.Atoi(hunkHeader[2])
+		nLines, _ := strconv.Atoi(hunkHeader[3])
+		hunk := &Hunk{
+			CommitID:  api.CommitID(commitID),
+			StartLine: lineNoCur,
+			EndLine:   lineNoCur + nLines,
+			StartByte: byteOffset,
+		}
+
+		if _, in := commits[commitID]; in {
+			// Already seen commit
+			byteOffset += len(remainingLines[1])
+			remainingLines = remainingLines[2:]
+		} else {
+			// New commit
+			author := strings.Join(strings.Split(remainingLines[1], " ")[1:], " ")
+			email := strings.Join(strings.Split(remainingLines[2], " ")[1:], " ")
+			if len(email) >= 2 && email[0] == '<' && email[len(email)-1] == '>' {
+				email = email[1 : len(email)-1]
+			}
+			authorTime, err := strconv.ParseInt(strings.Join(strings.Split(remainingLines[3], " ")[1:], " "), 10, 64)
+			if err != nil {
+				return nil, errors.Errorf("Failed to parse author-time %q", remainingLines[3])
+			}
+			summary := strings.Join(strings.Split(remainingLines[9], " ")[1:], " ")
+			commit := gitdomain.Commit{
+				ID:      api.CommitID(commitID),
+				Message: gitdomain.Message(summary),
+				Author: gitdomain.Signature{
+					Name:  author,
+					Email: email,
+					Date:  time.Unix(authorTime, 0).UTC(),
+				},
+			}
+
+			for i := 10; i < 13 && i < len(remainingLines); i++ {
+				if strings.HasPrefix(remainingLines[i], "filename ") {
+					filenames[commitID] = strings.SplitN(remainingLines[i], " ", 2)[1]
+					break
+				}
+			}
+
+			if len(remainingLines) >= 13 && strings.HasPrefix(remainingLines[10], "previous ") {
+				byteOffset += len(remainingLines[12])
+				remainingLines = remainingLines[13:]
+			} else if len(remainingLines) >= 13 && remainingLines[10] == "boundary" {
+				byteOffset += len(remainingLines[12])
+				remainingLines = remainingLines[13:]
+			} else if len(remainingLines) >= 12 {
+				byteOffset += len(remainingLines[11])
+				remainingLines = remainingLines[12:]
+			} else if len(remainingLines) == 11 {
+				// Empty file
+				remainingLines = remainingLines[11:]
+			} else {
+				return nil, errors.Errorf("Unexpected number of remaining lines (%d):\n%s", len(remainingLines), "  "+strings.Join(remainingLines, "\n  "))
+			}
+
+			commits[commitID] = commit
+		}
+
+		if commit, present := commits[commitID]; present {
+			// Should always be present, but check just to avoid
+			// panicking in case of a (somewhat likely) bug in our
+			// git-blame parser above.
+			hunk.CommitID = commit.ID
+			hunk.Author = commit.Author
+			hunk.Message = string(commit.Message)
+		}
+
+		if filename, present := filenames[commitID]; present {
+			hunk.Filename = filename
+		}
+
+		// Consume remaining lines in hunk
+		for i := 1; i < nLines; i++ {
+			byteOffset += len(remainingLines[1])
+			remainingLines = remainingLines[2:]
+		}
+
+		hunk.EndByte = byteOffset
+		hunks = append(hunks, hunk)
+	}
+
+	return hunks, nil
+}
+
+func gitserverGitCommandFunc(repo api.RepoName, db database.DB) gitCommandFunc {
+	return func(args []string) GitCommand {
+		return NewClient(db).GitCommand(repo, args...)
+	}
+}
+
+// gitCommandFunc is a func that creates a new executable Git command.
+type gitCommandFunc func(args []string) GitCommand
+
+// IsAbsoluteRevision checks if the revision is a git OID SHA string.
+//
+// Note: This doesn't mean the SHA exists in a repository, nor does it mean it
+// isn't a ref. Git allows 40-char hexadecimal strings to be references.
+func IsAbsoluteRevision(s string) bool {
+	if len(s) != 40 {
+		return false
+	}
+	for _, r := range s {
+		if !(('0' <= r && r <= '9') ||
+			('a' <= r && r <= 'f') ||
+			('A' <= r && r <= 'F')) {
+			return false
+		}
+	}
+	return true
+}
+
+// ResolveRevisionOptions configure how we resolve revisions.
+// The zero value should contain appropriate default values.
+type ResolveRevisionOptions struct {
+	NoEnsureRevision bool // do not try to fetch from remote if revision doesn't exist locally
+}
+
+var resolveRevisionCounter = promauto.NewCounterVec(prometheus.CounterOpts{
+	Name: "src_resolve_revision_total",
+	Help: "The number of times we call internal/vcs/git/ResolveRevision",
+}, []string{"ensure_revision"})
+
+// ResolveRevision will return the absolute commit for a commit-ish spec. If spec is empty, HEAD is
+// used.
+//
+// Error cases:
+// * Repo does not exist: gitdomain.RepoNotExistError
+// * Commit does not exist: gitdomain.RevisionNotFoundError
+// * Empty repository: gitdomain.RevisionNotFoundError
+// * Other unexpected errors.
+func ResolveRevision(ctx context.Context, db database.DB, repo api.RepoName, spec string, opt ResolveRevisionOptions) (api.CommitID, error) {
+	if Mocks.ResolveRevision != nil {
+		return Mocks.ResolveRevision(spec, opt)
+	}
+
+	labelEnsureRevisionValue := "true"
+	if opt.NoEnsureRevision {
+		labelEnsureRevisionValue = "false"
+	}
+	resolveRevisionCounter.WithLabelValues(labelEnsureRevisionValue).Inc()
+
+	span, ctx := ot.StartSpanFromContext(ctx, "Git: ResolveRevision")
+	span.SetTag("Spec", spec)
+	span.SetTag("Opt", fmt.Sprintf("%+v", opt))
+	defer span.Finish()
+
+	if err := checkSpecArgSafety(spec); err != nil {
+		return "", err
+	}
+	if spec == "" {
+		spec = "HEAD"
+	}
+	if spec != "HEAD" {
+		// "git rev-parse HEAD^0" is slower than "git rev-parse HEAD"
+		// since it checks that the resolved git object exists. We can
+		// assume it exists for HEAD, but for other commits we should
+		// check.
+		spec = spec + "^0"
+	}
+
+	cmd := NewClient(db).GitCommand(repo, "rev-parse", spec)
+	cmd.SetEnsureRevision(spec)
+
+	// We don't ever need to ensure that HEAD is in git-server.
+	// HEAD is always there once a repo is cloned
+	// (except empty repos, but we don't need to ensure revision on those).
+	if opt.NoEnsureRevision || spec == "HEAD" {
+		cmd.SetEnsureRevision("")
+	}
+
+	return runRevParse(ctx, cmd, spec)
+}
+
+// runRevParse sends the git rev-parse command to gitserver. It interprets
+// missing revision responses and converts them into RevisionNotFoundError.
+func runRevParse(ctx context.Context, cmd GitCommand, spec string) (api.CommitID, error) {
+	stdout, stderr, err := cmd.DividedOutput(ctx)
+	if err != nil {
+		if gitdomain.IsRepoNotExist(err) {
+			return "", err
+		}
+		if bytes.Contains(stderr, []byte("unknown revision")) {
+			return "", &gitdomain.RevisionNotFoundError{Repo: cmd.Repo(), Spec: spec}
+		}
+		return "", errors.WithMessage(err, fmt.Sprintf("git command %v failed (stderr: %q)", cmd.Args(), stderr))
+	}
+	commit := api.CommitID(bytes.TrimSpace(stdout))
+	if !IsAbsoluteRevision(string(commit)) {
+		if commit == "HEAD" {
+			// We don't verify the existence of HEAD (see above comments), but
+			// if HEAD doesn't point to anything git just returns `HEAD` as the
+			// output of rev-parse. An example where this occurs is an empty
+			// repository.
+			return "", &gitdomain.RevisionNotFoundError{Repo: cmd.Repo(), Spec: spec}
+		}
+		return "", gitdomain.BadCommitError{Spec: spec, Commit: commit, Repo: cmd.Repo()}
+	}
+	return commit, nil
 }
