+++ internal/gitserver/mocks.go
@@ -17,8 +17,9 @@ import (
 // that's done, we should take advantage of the generated mock client in this
 // package instead.
 var Mocks, emptyMocks struct {
-	ExecReader func(args []string) (reader io.ReadCloser, err error)
-	ReadDir    func(commit api.CommitID, name string, recurse bool) ([]fs.FileInfo, error)
+	ExecReader      func(args []string) (reader io.ReadCloser, err error)
+	ReadDir         func(commit api.CommitID, name string, recurse bool) ([]fs.FileInfo, error)
+	ResolveRevision func(spec string, opt ResolveRevisionOptions) (api.CommitID, error)
 }
 
 // ResetMocks clears the mock functions set on Mocks (so that subsequent tests don't inadvertently
